requirements.txt
ansible
linode_api4
dnspython
python-gitlab
boto3
hvac
prettytable
jmespath
PyNaCl
